delete: hardhat/evm - Support around custom verifier, signer, contracts and tasks (pulled from BalancerV2 Monorepo) - Clean up imports in hardhat.config.ts and move networks there
